+++ src/layouts/BasicLayout.js
@@ -7,10 +7,10 @@ import { ContainerQuery } from 'react-container-query';
 import classNames from 'classnames';
 import pathToRegexp from 'path-to-regexp';
 import { formatMessage } from 'umi/locale';
-import SiderMenu from '../../components/SiderMenu';
-import Authorized from '../../utils/Authorized';
-import SettingDarwer from '../../components/SettingDarwer';
-import logo from '../../assets/logo.svg';
+import SiderMenu from '../components/SiderMenu/index';
+import Authorized from '../utils/Authorized';
+import SettingDarwer from '../components/SettingDarwer/index';
+import logo from '../assets/logo.svg';
 import Footer from './Footer';
 import Header from './Header';
 import Context from './MenuContext';
