+++ src/layouts/Header.js
@@ -3,10 +3,10 @@ import { Layout, message } from 'antd';
 import Animate from 'rc-animate';
 import { connect } from 'dva';
 import { routerRedux } from 'dva/router';
-import GlobalHeader from '../../components/GlobalHeader';
-import TopNavHeader from '../../components/TopNavHeader';
+import GlobalHeader from '../components/GlobalHeader/index';
+import TopNavHeader from '../components/TopNavHeader/index';
 import styles from './Header.less';
-import Authorized from '../../utils/Authorized';
+import Authorized from '../utils/Authorized';
 
 const { Header } = Layout;
