+++ src/pages/Forms/AdvancedForm.js
@@ -14,7 +14,7 @@ import {
 } from 'antd';
 import { connect } from 'dva';
 import FooterToolbar from 'components/FooterToolbar';
-import PageHeaderLayout from '../layouts/PageHeaderLayout';
+import PageHeaderLayout from '../../layouts/PageHeaderLayout';
 import TableForm from './TableForm';
 import styles from './style.less';
