+++ src/pages/List/BasicList.js
@@ -21,7 +21,7 @@ import {
   Select,
 } from 'antd';
 
-import PageHeaderLayout from '../layouts/PageHeaderLayout';
+import PageHeaderLayout from '../../layouts/PageHeaderLayout';
 import Result from '../../components/Result';
 
 import styles from './BasicList.less';
