+++ src/pages/List/CardList.js
@@ -3,7 +3,7 @@ import { connect } from 'dva';
 import { Card, Button, Icon, List } from 'antd';
 
 import Ellipsis from 'components/Ellipsis';
-import PageHeaderLayout from '../layouts/PageHeaderLayout';
+import PageHeaderLayout from '../../layouts/PageHeaderLayout';
 
 import styles from './CardList.less';
