+++ src/pages/Profile/AdvancedProfile.js
@@ -19,7 +19,7 @@ import {
 } from 'antd';
 import classNames from 'classnames';
 import DescriptionList from 'components/DescriptionList';
-import PageHeaderLayout from '../layouts/PageHeaderLayout';
+import PageHeaderLayout from '../../layouts/PageHeaderLayout';
 import styles from './AdvancedProfile.less';
 
 const { Step } = Steps;
